Refactor the class SlowRuntimeStub
Because there are many redundant codes between class SlowRuntimeStub and
class RuntimeStubs, the shared logic should be extracted. Notice that
the primary difference between both stubs is the types of arguments, the
methods of RuntimeStubs accept JSHandles while that of SlowRuntimeStub
are JSTaggedValue. This PR removes the redundant implementations of
class SlowRuntimeStubs via invoking the methods from RuntimStubs, in
this way, the program is easy to maintain and read in the future.

Testing: pass all the CI tests, because there are not any functional
changes.

Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5J59X

Signed-off-by: wuzhefeng <wuzhefeng1@huawei.com>
Change-Id: I556e88a8ae01edb72fc761ea5a674514ee470109