Fix the bug when adding elements to TSObjLayoutInfo
A new bug was introduced recently. When adding elements to TSObjLayoutInfo sequentially,
the location of the addition should be consistent with the number of times it is added
instead of the outer loop's index, otherwise the location will violate the assertion
and cause the program to crash. Tracing back to the root cause,
the design of TSObjLayoutInfo::SetKeyAndType is confusing.
The index parameter is used to validate the position,
not to indicate the location of the setting. We need to fix it.

Issue:
https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I6HELT

Tests:
typeinfer test, aot test

Signed-off-by: huoqingyi <huoqingyi@huawei.com>
Change-Id: Idb7537016431569d16b95009cc48915944d9f208