Adjust the number of types that GlobalTSTypeRef can represent
1. During parsing TSTypes, when parsing to a type that reachs the maximum number of
types that GlobalTSTypeRef can represent, the program should continue to execute
instead of reporting an error. We need to fix it.
2. At the same time, we also checked the entire parsing, in some cases,
the failure of parsing should not cause the program to crash.
3. In the process of modification, we found that the class TSManager was already large,
so we add a new class to handle the new scenario about TSTypeTable.
Also we move some interfaces of TSManager to this new class.

Issue:
https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I6JCNA

Tests:
typeinfer test, aot test, TSTypeTest

Change-Id: Ia18828eb58affd2bf017436b79aa7087cde24a2c
Signed-off-by: huoqingyi <huoqingyi@huawei.com>